feat(core): allow changing db params(only apply for dummy & read-only) (#398)
User can alter the params of read-only dummy storage engine. The HAS_DATA & IS_READ_ONLY attributes cannot be modified.

SQL
ALTER STORAGEENGINE <engine_id> WITH PARAMS "(<param_key>:<param_value>,?)+";
example:
alter storageengine 2 with params "dummy_dir:test/new_data";

——简化实现，后续随节点移除、容错能力的增强，需要相应调整实现

同时，按需相应调整测试脚本目录结构

+++ core/src/main/java/cn/edu/tsinghua/iginx/metadata/storage/etcd/ETCDMetaStorage.java
@@ -121,7 +121,7 @@ public class ETCDMetaStorage implements IMetaStorage {
   private final int STORAGE_ENGINE_NODE_LENGTH = 10;
 
   private String generateID(String prefix, long idLength, long val) {
-    return String.format(prefix + "%0" + idLength + "d", (int) val);
+    return String.format(prefix + "%0" + idLength + "d", val);
   }
 
   public ETCDMetaStorage() {
